I fixed syntax errors,and changed all api request files,moving them to a single file ApiRequest.js
