+++ src/Components/Footer_component.js
@@ -1,16 +1,17 @@
 import "../App.css";
 import React from "react";
 
-function Footer (){
-        return (
-            <footer className="App-Footer-background">
-                <p>Author: Pitter Parker</p>
-                <a
-                    href={"https://marvelcinematicuniverse.fandom.com/ru/wiki/%D0%A7%D0%B5%D0%BB%D0%BE%D0%B2%D0%B5%D0%BA-%D0%BF%D0%B0%D1%83%D0%BA"} className="App-Footer">
-                    Visit the site
-                </a>
-            </footer>
-        )
-    };
+function Footer() {
+    return (
+        <footer className="App-Footer-background">
+            <p>Author: Pitter Parker</p>
+            <p
+                href={"https://marvelcinematicuniverse.fandom.com/ru/wiki/%D0%A7%D0%B5%D0%BB%D0%BE%D0%B2%D0%B5%D0%BA-%D0%BF%D0%B0%D1%83%D0%BA"}
+                className="App-Footer">
+                Visit the site
+            </p>
+        </footer>
+    )
+};
 export default Footer;
