+++ src/Pages/Error404.js
@@ -8,4 +8,5 @@ function Error() {
         </header>
     );
 }
+
 export default Error;
